==> Part1 - MySQL database on Movies from a subset of IMDB's publicly available dataset.ipynb <==
# TMBD Movie Database Project - Part 1
- Scotty Thomason

## MySQL database on Movies from a subset of IMDB's publicly available dataset

import pandas as pd
import numpy as np
import os,json

## Akas DataFrame

# akas_url = "https://datasets.imdbws.com/title.akas.tsv.gz"
# akas = pd.read_csv(akas_url, sep='\t', low_memory=False)
# akas.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)

akas = pd.read_csv("Data/title_akas.csv.gz", low_memory = False)

akas.head()

akas.replace({'\\N':np.nan}, inplace = True)

us_mask = (akas['region'] == 'US')
filter_akas = akas.loc[us_mask]

filter_akas.head()

filter_akas.info()

## Save current dataframe to file.
# filter_akas.to_csv("Data/filtered_akas.csv.gz",compression='gzip',index=False)

## Basics DataFrame

#Data Sets
# basics_url="https://datasets.imdbws.com/title.basics.tsv.gz"

# Loading in the Data and saving it locally
# basics = pd.read_csv(basics_url, sep='\t', low_memory=False)
# basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)
basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)

# Filter the basics table down to only include the US by using the filter akas dataframe
keepers = basics['tconst'].isin(filter_akas['titleId'])
keepers

basics = basics[keepers]
basics.head()

basics.info()

basics.replace({'\\N':np.nan}, inplace = True)

# Eliminate movies that are null for runtimeMinutes
basics.dropna(subset=['runtimeMinutes'], inplace = True)
basics.dropna(subset=['genres'], inplace = True)

basics['titleType'].value_counts()

mask1 = (basics['titleType'] == 'movie')
filter_basics = basics.loc[mask1]

mask2 = (filter_basics['startYear'] >= '2000') & (filter_basics['startYear'] < '2022')
filter_basics1 = filter_basics.loc[mask2]

filter_basics1.head()

# Exclude movies that are included in the documentary category.
is_documentary = filter_basics1['genres'].str.contains('documentary',case=False)
filter_basics2 = filter_basics1[~is_documentary]

filter_basics2.head()

filter_basics2.info()

## Save current dataframe to file.
# filter_basics2.to_csv("Data/filtered_basics.csv.gz",compression='gzip',index=False)

## Third DataFrame

# ratings_url = "https://datasets.imdbws.com/title.ratings.tsv.gz"
# ratings = pd.read_csv(ratings_url, sep='\t', low_memory=False)
# ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)
ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)

ratings.head()

ratings.replace({'\\N':np.nan}, inplace = True)

keepers = ratings['tconst'].isin(filter_akas['titleId'])
ratings = ratings[keepers]

## Save current dataframe to file.
# ratings.to_csv("Data/filtered_ratings.csv.gz",compression='gzip',index=False)

ratings.info()
