==> Part 3 - Construct and export a MySQL database.ipynb <==
## Part 3 - Construct and export a MySQL database

# Normalizing Genres

import pandas as pd
import numpy as np
import os,json

basics_df = pd.read_csv('Data/filtered_basics.csv.gz', low_memory = False)
basics_df.head()

basics_df.drop(['titleType', 'originalTitle', 'isAdult'], axis=1, inplace=True)

## create a col with a list of genres
basics_df['genres_split'] = basics_df['genres'].str.split(',')
basics_df

basics_df.drop(columns= 'genres', inplace = True)

exploded_genres = basics_df.explode('genres_split')
exploded_genres

unique_genres = sorted(exploded_genres['genres_split'].unique())

title_genres = exploded_genres[['tconst', 'genres_split']].copy()
title_genres.head()

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

## make new integer genre_id and drop string genres
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres = title_genres.drop(columns='genres_split')

basics_df.drop(columns = 'genres_split', inplace = True)

genre_lookup = pd.DataFrame({'genre_name': genre_map.keys(), 'genre_id': genre_map.values()})
genre_lookup.head()

# Saving the MySQL tables with tconst as the primary key.

import pandas as pd
from sqlalchemy_utils import database_exists, create_database
import pymysql
from sqlalchemy import create_engine
from sqlalchemy.types import *
pymysql.install_as_MySQLdb()

with open('/Users/scyjt/.secret/mysql.json') as f:
    login = json.load(f)
login.keys()

from urllib.parse import quote_plus as urlquote
connection = f"mysql+pymysql://{login['username']}:{urlquote(login['password'])}@localhost/movies"
engine = create_engine(connection)

if database_exists(connection):
    print('It exists!')
else:
    create_database(connection)
    print('Database Created')

## Calculate max string lengths for object columns
key_len = basics_df['tconst'].fillna('').map(len).max()
title_len = basics_df['primaryTitle'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Float(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}



# Save to sql with dtype and index=False
basics_df.to_sql('title_basics',engine,dtype=df_schema,if_exists='replace',index=False)


title_genres.to_sql('title_genres',engine,if_exists='replace',index=False)

genre_lookup.to_sql('genres',engine,if_exists='replace',index=False)

## Create a schema dictonary using Sqlalchemy datatype objects
df_schema_ratings = {
    "tconst": String(key_len+1), 
    "averageRating": Text(title_len+1),
    'numVotes':Float()}

title_ratings = pd.read_csv('Data/filtered_ratings.csv.gz', low_memory = False)

title_ratings.to_sql('title_ratings',engine,dtype=df_schema_ratings,if_exists='replace',index=False)

## Create a schema dictonary using Sqlalchemy datatype objects
df_schema_tmdb = {
    "tconst": String(key_len+1), 
    'revenue':Float(),
    'budget':Float(),
    'certification':Integer()}

tmdb_data = pd.read_csv('Data/tmdb_results_combined.csv.gz', low_memory = False)

tmdb_data.to_sql('tmdb_data',engine,dtype=df_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')
